Always switch logouhr on when space is opened
Change the default behaviour to "always on when somebody is there".
Normally we want to see this device, so it may as well be activated.
Deactivation can be achieved via OpenHAB control panel or some other
means (like a script on Konfuzius).

Makes us more independent from the double-tap on a sonoff device.